requirements.txt
llama-index>=0.10.0
anthropic>=0.6.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-socketio>=5.3.0
llama-index-embeddings-huggingface